Move types to model dirs.
