+++ ui/src/app/base/reducers/pod/pod.ts
@@ -1,6 +1,6 @@
 import { createNextState } from "@reduxjs/toolkit";
 
-import { Pod } from "app/base/types";
+import { Pod } from "app/store/pod/types";
 
 const initialState = {
   errors: {},
