+++ ui/src/app/base/selectors/resourcepool/resourcepool.ts
@@ -1,6 +1,8 @@
 import { createSelector } from "@reduxjs/toolkit";
 
-import { ResourcePool, RootState, TSFixMe } from "app/base/types";
+import { ResourcePool } from "app/store/resourcepool/types";
+import { RootState } from "app/store/root/types";
+import { TSFixMe } from "app/base/types";
 
 /**
  * Returns all resource pools.
