+++ ui/src/app/kvm/views/KVMList/KVMListTable/NameColumn/NameColumn.tsx
@@ -4,7 +4,8 @@ import { useSelector } from "react-redux";
 import { Link } from "react-router-dom";
 
 import { pod as podSelectors } from "app/base/selectors";
-import { Pod, RootState } from "app/base/types";
+import { Pod } from "app/store/pod/types";
+import { RootState } from "app/store/root/types";
 import DoubleRow from "app/base/components/DoubleRow";
 
 type Props = {
