+++ ui/src/app/kvm/views/KVMList/KVMListTable/OSColumn/OSColumn.tsx
@@ -8,7 +8,7 @@ import {
   machine as machineSelectors,
   pod as podSelectors,
 } from "app/base/selectors";
-import { RootState } from "app/base/types";
+import { RootState } from "app/store/root/types";
 import DoubleRow from "app/base/components/DoubleRow";
 
 type Props = { id: number };
