+++ ui/src/app/kvm/views/KVMList/KVMListTable/VMsColumn/VMsColumn.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useSelector } from "react-redux";
 
 import { pod as podSelectors } from "app/base/selectors";
-import { RootState } from "app/base/types";
+import { RootState } from "app/store/root/types";
 import DoubleRow from "app/base/components/DoubleRow";
 
 type Props = { id: number };
