+++ ui/src/app/store/host/types.ts
@@ -1,0 +1,4 @@
+import type { Controller } from "app/store/controller/types";
+import type { Machine } from "app/store/machine/types";
+
+export type Host = Machine | Controller;
