+++ ui/src/app/store/types.ts
@@ -1,0 +1,3 @@
+export type Model = {
+  id: number;
+};
