+++ ui/src/app/store/utils.ts
@@ -1,0 +1,9 @@
+import type { Host } from "app/store/host/types";
+import type { Machine } from "app/store/machine/types";
+
+/**
+ * Type guard to determine if host is a machine.
+ * @param {Host} host - a machine or controller.
+ */
+export const isMachine = (host: Host): host is Machine =>
+  (host as Machine).link_type === "machine";
